Remove http_api.js from http_api/helpers/
http_api.js is redundant now on HTTP API module.
It also contains logic that belongs to init_steps/bootstrap_swagger.js
so this logic has been moved there.
There is only one function left: calcualteApproval that actually
should only belong to Chain module and would need to be exposed through
the channel as an action.
This can be developed further in another issue. For now, http_api.js
has been renamed to utils.js containing only calcualateApproval function.
