+++ framework/src/modules/chain/helpers/index.js
@@ -5,7 +5,6 @@
  * @property {module:helpers/diff} Diff
  * @property {module:helpers/ed} Ed
  * @property {module:helpers/git} Git
- * @property {module:helpers/http_api} HTTP_API
  * @property {module:helpers/jobs_queue} Jobs_queue
  * @property {module:helpers/slots} Slots
  * @property {module:helpers/sort_by} Sort_by
